requirement.txt
SQLAlchemy==1.4.22
psycopg2==2.9.1
dask
python-dotenv==0.19.0